pyproject.toml
[tool.poetry]
license = "MIT"
name = "project_name"
version = "0.1.0"
description = "A service that gets a webpage and extracts meta data"
readme = "README.md"
authors = ["Fabrizio Bonavita <fabrizioschmidt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
pydantic-sqlalchemy = "^0.0.8"
alembic = "^1.5.8"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
coverage = {extras = ["toml"], version = "^5.3"}
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.9.0"
mypy = "^0.790"
flake8-annotations = "^2.4.1"
typeguard = "^2.10.0"
flake8-docstrings = "^1.5.0"
darglint = "^1.5.6"
Sphinx = "^3.3.1"
xdoctest = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["project_name"]

[tool.coverage.report]
show_missing = true
fail_under = 100
